requirements.txt
fastapi
uvicorn
gspread
oauth2client
python-dotenv
sqlalchemy